move ZookeeperPool to util.zk
