+++ paasta_tools/autoscaling/autoscaling_service_lib.py
@@ -65,11 +65,11 @@ from paasta_tools.util.config_loading import list_services
 from paasta_tools.util.config_loading import load_system_paasta_config
 from paasta_tools.util.config_loading import SystemPaastaConfig
 from paasta_tools.util.const import DEFAULT_SOA_DIR
+from paasta_tools.util.zk import ZookeeperPool
 from paasta_tools.utils import _log
 from paasta_tools.utils import get_user_agent
 from paasta_tools.utils import mean
 from paasta_tools.utils import use_requests_cache
-from paasta_tools.utils import ZookeeperPool
 
 try:
     import yelp_meteorite
