+++ paasta_tools/autoscaling/load_boost.py
@@ -18,7 +18,7 @@ from kazoo.exceptions import NoNodeError
 
 from paasta_tools.clusterman import get_clusterman_metrics
 from paasta_tools.util.config_loading import load_system_paasta_config
-from paasta_tools.utils import ZookeeperPool
+from paasta_tools.util.zk import ZookeeperPool
 
 clusterman_metrics, __ = get_clusterman_metrics()
 DEFAULT_BOOST_FACTOR = 1.5
