+++ paasta_tools/deployd/master.py
@@ -30,7 +30,7 @@ from paasta_tools.marathon_tools import DEFAULT_SOA_DIR
 from paasta_tools.metrics.metrics_lib import get_metrics_interface
 from paasta_tools.util.config_loading import get_services_for_cluster
 from paasta_tools.util.config_loading import load_system_paasta_config
-from paasta_tools.utils import ZookeeperPool
+from paasta_tools.util.zk import ZookeeperPool
 
 # Broken out into a constant so that we don't get drift between this and the code in paasta_deployd_steps.py that
 # searches for this message.
