+++ tests/test_firewall_update.py
@@ -66,7 +66,7 @@ def test_parse_args_default_cron():
 
 
 @mock.patch(
-    "paasta_tools.utils.load_system_paasta_config", autospec=True,
+    "paasta_tools.util.config_loading.load_system_paasta_config", autospec=True,
 )
 @mock.patch.object(
     firewall_update,
